test(fresco): split frame tests to respect source length limit
